% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tristan-package.R
\name{augment_posterior}
\alias{augment_posterior}
\alias{augment_posterior_predict}
\alias{augment_posterior_linpred}
\title{Create a long data-frame of posterior predictions}
\usage{
augment_posterior_predict(model, newdata = NULL, ..., nsamples = NULL)

augment_posterior_linpred(model, newdata = NULL, ..., nsamples = NULL)
}
\arguments{
\item{model}{a model fit with \link[rstanarm:rstanarm]{RStanARM}}

\item{newdata}{a dataframe of new observations to generate predictions. If
omitted or null, the original data-set used to fit the model is used for
the predictions.}

\item{...}{other arguments passed to \code{\link[rstanarm:posterior_predict]{rstanarm::posterior_predict()}} or
\code{\link[rstanarm:posterior_linpred]{rstanarm::posterior_linpred()}}.}

\item{nsamples}{number of samples to draw from the posterior. Defaults to the
full number of samples.}
}
\value{
a data-frame (a \code{\link[tibble:tibble]{tibble::tibble()}}) of posterior predictions. The
dataset used to generate the predictions is augmented with posterior
predictions, so that there is row per posterior prediction per observation.
Additionaly columns are included: \code{.observation} uniquely identifies rows
from the prediction data-set, \code{.draw} is the posterior sample number, and
\code{.posterior_value} is the predicted value for that \code{.observation} in that
\code{.draw}.
}
\description{
These functions are wrapper around RStanARM's prediction functions that
return a long tidy dataframe of model predictions.
\code{augment_posterior_predict()} wraps \code{\link[rstanarm:posterior_predict]{rstanarm::posterior_predict()}} and
\code{augment_posterior_linpred()} wraps \code{\link[rstanarm:posterior_predict]{rstanarm::posterior_predict()}}.
}
