% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tristan-package.R
\name{augment_posterior_linpred}
\alias{augment_posterior_linpred}
\title{Create a long data-frame of linear-posterior predictions}
\usage{
augment_posterior_linpred(model, newdata = NULL, ..., nsamples = NULL)
}
\description{
Create a long data-frame of linear-posterior predictions
}
