% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tristan-package.R
\name{augment_posterior_predict}
\alias{augment_posterior_predict}
\title{Create a long data-frame of posterior predictions}
\usage{
augment_posterior_predict(model, newdata = NULL, ..., nsamples = NULL)
}
\description{
Create a long data-frame of posterior predictions
}
