% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2.R
\name{calculate_model_r2}
\alias{calculate_model_r2}
\title{Calculate classical r-squared from a model}
\usage{
calculate_model_r2(model)
}
\arguments{
\item{model}{an RStanARM model}
}
\value{
the rsquared value for each posterior sample
}
\description{
Calculate classical r-squared from a model
}
