% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2.R
\name{calculate_r2}
\alias{calculate_r2}
\title{Calculate classical r-squared from observations and predictions}
\usage{
calculate_r2(y_obs, y_fits)
}
\arguments{
\item{y_obs}{observed y values}

\item{y_fits}{fitted y values}
}
\value{
the rsquared value
}
\description{
Calculate classical r-squared from observations and predictions
}
