% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervals.R
\name{double_etdi}
\alias{double_etdi}
\title{Calculate two ETDI intervals}
\usage{
double_etdi(x, prob_outer, prob_inner)
}
\arguments{
\item{x}{a model fitted with RStanARM or a vector of MCMC samples}

\item{prob_outer, prob_inner}{width of probability intervals}
}
\value{
a data-frame (a \code{\link[tibble:tibble]{tibble::tibble()}}) of intervals. Contains columns
with the \code{term} name, \code{outer_density}/\code{inner_density}, and the
\code{outer_lower}/\code{inner_lower} and \code{inner_upper}/\code{outer_upper} values of the
interval.
}
\description{
This function helps create caterpillar plots.
}
