% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervals.R
\name{posterior-intervals}
\alias{posterior-intervals}
\alias{tidy_hpdi}
\alias{tidy_hpdi.numeric}
\alias{tidy_hpdi.stanreg}
\alias{tidy_hpdi.mcmc.list}
\alias{tidy_etdi}
\alias{tidy_etdi.numeric}
\alias{tidy_etdi.stanreg}
\alias{tidy_etdi.mcmc.list}
\title{Calculate posterior density intervals}
\usage{
tidy_hpdi(x, prob = 0.9)

\method{tidy_hpdi}{numeric}(x, prob = 0.9)

\method{tidy_hpdi}{stanreg}(x, prob = 0.9)

\method{tidy_hpdi}{mcmc.list}(x, prob = 0.9)

tidy_etdi(x, prob = 0.9)

\method{tidy_etdi}{numeric}(x, prob = 0.9)

\method{tidy_etdi}{stanreg}(x, prob = 0.9)

\method{tidy_etdi}{mcmc.list}(x, prob = 0.9)
}
\arguments{
\item{x}{a model fitted with RStanARM or a vector of MCMC samples}

\item{prob}{a probability content or width of the interval. Defaults to .9.}
}
\value{
a data-frame (a \code{\link[tibble:tibble]{tibble::tibble()}}) of intervals. Contains columns
with the \code{term} name, \code{interval} type and \code{density}, and the \code{lower}
and \code{upper} values of the interval.
}
\description{
Calculate the highest posterior density interval (HPDI) or the equal-tailed
density interval (ETDI) of posterior samples.
}
\details{
The equal-tailed density interval is a term I made up for the traditional
sort of confidence or uncertainty interval. A 90% ETDI tells you the value at
the 5th and 95th percentiles. The interval contains 90% of the probability
density, and the left and right tails outside of the interval contain an
equal amount of probability density (5%). I use \code{\link[stats:quantile]{stats::quantile()}} to
compute the interval.

The highest density posterior interval is the narrowest interval with a given
probability density. A 90% HPDI tells you the skinniest interval that
contains 90% of the probability density. The left and right tails outside of
the interval might contain unequal amounts of probability density. I use
\code{\link[coda:HPDinterval]{coda::HPDinterval()}} to compute these intervals.
}
