% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classical.R
\name{stan_to_classical}
\alias{stan_to_classical}
\alias{stan_to_glm}
\alias{stan_to_lm}
\title{Refit a simple RStanARM model with a classical version}
\usage{
stan_to_glm(model)

stan_to_lm(model)
}
\arguments{
\item{model}{a model fit with \code{\link[rstanarm:stan_glm]{rstanarm::stan_glm()}} or
\code{\link[rstanarm:stan_lm]{rstanarm::stan_lm()}}.}
}
\value{
the corresponding model fit with \code{\link[stats:glm]{stats::glm()}} or
\code{\link[stats:lm]{stats::lm()}}.
}
\description{
Refit a simple RStanARM model with a classical version
}
