% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervals.R
\name{tidy_median}
\alias{tidy_median}
\title{Calculate the median for posterior samples}
\usage{
tidy_median(x)
}
\arguments{
\item{x}{a model fitted with RStanARM or a vector of MCMC samples}
}
\value{
a data-frame (a \code{\link[tibble:tibble]{tibble::tibble()}}) with the median values.
}
\description{
Calculate the median for posterior samples
}
