% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tristan-package.R
\docType{package}
\name{tristan}
\alias{tristan}
\alias{tristan-package}
\title{Tristan's helper functions for RStanARM models and MCMC samples}
\description{
Tristan's helper functions for RStanARM models and MCMC samples
}
