% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tristan-package.R
\name{ggs_rstanarm}
\alias{ggs_rstanarm}
\title{Create tidy dataframe for ggmcmc from an RStanARM model}
\usage{
ggs_rstanarm(model)
}
\arguments{
\item{model}{a model fit with \link[rstanarm:rstanarm]{RStanARM}}
}
\value{
a data-frame (a \code{\link[tibble:tibble]{tibble::tibble()}}) of MCMC sampling information
suitable for use with the \link[ggmcmc:ggmcmc]{ggmcmc package}.
}
\description{
Create tidy dataframe for ggmcmc from an RStanARM model
}
